+++ nuclia/sdk/upload.py
@@ -236,7 +236,7 @@ class NucliaUpload:
             try:
                 resource = ndb.ndb.get_resource_by_slug(kbid=ndb.kbid, slug=slug)
                 rid = resource.id
-                logger.warn(f"Using existing resource: {rid}")
+                logger.warning(f"Using existing resource: {rid}")
                 need_to_create_resource = False
             except exceptions.NotFoundError:
                 need_to_create_resource = True
@@ -253,7 +253,7 @@ class NucliaUpload:
                     kw[param] = kwargs.get(param)
             resource = ndb.ndb.create_resource(**kw)
             rid = resource.uuid
-            logger.warn(f"New resource created: {rid}")
+            logger.warning(f"New resource created: {rid}")
 
         return (rid, need_to_create_resource)