+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## 1.1.0 (unreleased)
 
-- Default output in JSON, + support `--indent` and `--yaml` options
+- Support `--json` and `--yaml` options
 - Rename `ask` in `chat` as `ask` will be another new feature
+- Support all regular Nuclia resource metadata
+- Update the resource if it exists
 
 ## 1.0.4 (2023-06-29)